pkp/pkp-lib#12104 changes after review
